FlaskRestApi/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
autopep8==1.6.0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
Werkzeug==2.0.2
zipp==3.7.0
